+++ src/main/java/com/application/internshipbackend/payload/request/AuthenticationRequest.java
@@ -14,5 +14,5 @@ public class AuthenticationRequest implements UserEmailRequest{
     private String password;
 
     @Builder.Default
-    private Integer deviceId = null;
+    private Long deviceId = null;
 }
